requirements.txt
tk
numpy
Pillow
pyequilib
torch